Update query.sparql
This change removes the bulk of false positives and gives ~8000 matches for english. There is still room for improvement, but this is a good start that won't drive the player (me!) crazy.

Note I had to change the language option to only return one language at a time because it times out otherwise. Can you run the query once for each language?