Increase maxlag for saving results in Wikidata
maxlag = 5 is the recommended value for bots and therefore the default
in LexData. Since this is not a bot but an interactive application, we
can use a significantly higher value. The value is chosen somewhat
arbitrary from the common value of maxlag in a way that is should only
block in rare cases.